bump version 0.1.0
add panel for warning notifications

uses shadow DOM so css is seperate from the rest of the website.

This does mean that the css needs to be instead added below the shadow
root in either a stylesheet or style
tags or inline

i chose to use style tags for each component
which probably will need to be changed
to instead store them in a
single style tag as a direct
shadow root child

added typescript related css module
typing to stop errors
as the css files are not modules
they are not being used effectively

start using BEM for css names

basic panel and button demos added
don't have logic yet

remove unused print script

update package.json and eslintrc

- update eslint versio
- update eslint-config-react-app
- update eslint-plugin-react-hooks
- add type-fest for slightly nicer JsonValue type
-add mui material library for nice components

- add plugins for typescript and react
for eslint
  > typescript linting way better now
- allow destructuring to let but not
complain if one of the vars doesn't change

add vscode config to debug extension in chrome

can be slow but works in less than a minute or so

huge update

- move jwt reader and fail log from
popup and options respectively to
warning buttons injected on page

- add FunctionComponent types

- save JWTs as objects with header,
body,signature instead of a string
 * no reason to parse it twice
- Rename JWT to JWTMessage to
distinguish it from JWT which does
not include meta-data.

- rename Receiver to JWTTable
for clarity

- add warnings HUD elements
 - includes button and modal
 - both use material ui components

 - move message saving inside storageReader
 * now need to reference it's data property to access messages

- restrict chrome.storage keys to
be specific strings and not arbitrary

- add missing return statement to decodeRaw.decodedData

- remove default argument from save
to storage procedure